Add say command and add auto delete
This PR adds the ability to auto delete error messages as mentioned in #7 . The time to auto delete is set to 30000 ms(30 seconds), You can change it you want. This PR also adds ability to repeat the saying with a command. For example,

I say  `!say hello`
And the bot will say `hello`.

^^^^^^^